Rename families.py to kv_format_spec_families.py
